Registration, Login and other little aggiusti made on the businessprofile component (Cards)
Registrations in the json is now completed, a new user is going to get stored correctly after he has done the test. Login is now working with new user also after the reload of the page because it's stored in the json. For business profile page all the cards now have the counter, added a little css propr for the text-decoration of the <a> tags.