Changes on the button to get in the test, Add itinerary is using a new json, same for stageHallCard and the summaryItinerary
When u click on the button to get into the test after registration the page will not be reloaded so u can do it without problems. The stageHallCard is now getting data from Suggested.json file. Also methods i need to fix just the method for the summaryStage. In summary itinerary for the slider i've used a double v-for cycle but maybe the slick can't work with it