Edit profile page is now working as it Should
i've hadded a method to get the data of the user logged, to change this data and to save it after clicking the save button. After that the user will get on the business page or private page based on in Type of user. On the new log in the user has to use the new email