Started to implement the option to be a business user
When the user is in the private Profile page can click on the business button. When he clicks he can change and set some options. For now he can set those options and see them during the inSession by clicking on modify in the businesscard4. Need to fix some things and add the post method for json server. I've also changed some of the a hrefs cause where reloading all the time the page and some other things. Added some new values for the agency or payment method in the users.json.....