Business button is now working
i've added new methods to change the payment method and agency details. I've also changed some v-if and v-else condition for the check of payment method so if a user select 1 and then wants to change it he will click on modify and see the payment that he clicked before. Same thing is working for the agency details...